requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
langchain_core
streamlit